Refactor Model.from_config
