+++ src/graphnet/models/model.py
@@ -3,6 +3,7 @@
 from abc import ABC, abstractmethod
 import dill
 import os.path
+import re
 from typing import Dict, List, Optional, Union
 
 try:
@@ -19,8 +20,13 @@ import torch
 from torch import Tensor
 from torch_geometric.data import Data
 
+import graphnet
 from graphnet.utilities.logging import LoggerMixin
 from graphnet.utilities.config import Configurable, ModelConfig
+from graphnet.utilities.config.parsing import (
+    traverse_and_apply,
+    get_all_grapnet_classes,
+)
 
 
 class Model(Configurable, LightningModule, LoggerMixin, ABC):
@@ -92,4 +98,31 @@ class Model(Configurable, LightningModule, LoggerMixin, ABC):
         assert isinstance(
             source, ModelConfig
         ), f"Argument `source` of type ({type(source)}) is not a `ModelConfig"
-        return source.construct_model(trust=trust, load_modules=load_modules)
+
+        # Check(s)
+        if load_modules is None:
+            load_modules = ["torch"]
+        assert isinstance(load_modules, list)
+
+        # Get a lookup for all classes in `graphnet`
+        namespace_classes = get_all_grapnet_classes(
+            graphnet.data, graphnet.models, graphnet.training
+        )
+
+        # Load any additional modules into the global namespace
+        for module in load_modules:
+            assert re.match("^[a-zA-Z_]+$", module) is not None
+            if module in globals():
+                continue
+            exec(f"import {module}", globals())
+
+        # Parse potential ModelConfig arguments
+        arguments = dict(**source.arguments)
+        arguments = traverse_and_apply(
+            arguments,
+            source._deserialise,
+            fn_kwargs={"trust": trust},
+        )
+
+        # Construct model based on arguments
+        return namespace_classes[source.class_name](**arguments)
