+++ src/graphnet/utilities/config/model_config.py
@@ -1,49 +1,22 @@
 """Config classes for the `graphnet.models` module."""
 from functools import wraps
 import inspect
-import itertools
-import pkgutil
-import re
-import types
 from typing import (
     TYPE_CHECKING,
     Any,
     Callable,
     Dict,
-    List,
     Union,
-    Optional,
-    OrderedDict,
 )
 
-import graphnet
 from graphnet.utilities.config.base_config import (
     BaseConfig,
     get_all_argument_values,
 )
-from graphnet.utilities.logging import get_logger
+from graphnet.utilities.config.parsing import traverse_and_apply
 
 if TYPE_CHECKING:
     from graphnet.models import Model
-
-logger = get_logger()
-
-
-def traverse_and_apply(
-    obj: Any, fn: Callable, fn_kwargs: Optional[Dict[str, Any]] = None
-) -> Any:
-    """Apply `fn` to all elements in `obj`, resulting in same structure."""
-    if isinstance(obj, (list, tuple)):
-        return [traverse_and_apply(elem, fn, fn_kwargs) for elem in obj]
-    elif isinstance(obj, dict):
-        return {
-            key: traverse_and_apply(val, fn, fn_kwargs)
-            for key, val in obj.items()
-        }
-    else:
-        if fn_kwargs is None:
-            fn_kwargs = {}
-        return fn(obj, **fn_kwargs)
 
 
 class ModelConfig(BaseConfig):
@@ -75,9 +48,6 @@ class ModelConfig(BaseConfig):
 
             In another session, you can then do:
             >>> model = Model.from_config("model.yml")
-
-            Or identically:
-            >>> model = ModelConfig.load("model.yml").construct_model()
         """
         # Parse any nested `ModelConfig` arguments
         for arg in data["arguments"]:
@@ -94,50 +64,6 @@ class ModelConfig(BaseConfig):
 
         # Base class constructor
         super().__init__(**data)
-
-    def construct_model(
-        self, trust: bool = False, load_modules: Optional[List[str]] = None
-    ) -> "Model":
-        """Construct `Model` based on current config.
-
-        Arguments:
-            trust: Whether to trust the ModelConfig file enough to  `eval(...)`
-                any lambda function expressions contained.
-            load_modules: List of modules used in the definition of the model
-                which, as a consequence, need to be loaded into the global
-                namespace. Defaults to loading `torch`.
-
-        Raises:
-            ValueError: If the ModelConfig contains lambda functions but
-                `trust = False`.
-        """
-        # Check(s)
-        if load_modules is None:
-            load_modules = ["torch"]
-
-        # Get a lookup for all classes in `graphnet`
-        namespace_classes = get_all_grapnet_classes(
-            graphnet.data, graphnet.models, graphnet.training
-        )
-
-        # Load any additional modules into the global namespace
-        if load_modules:
-            for module in load_modules:
-                assert re.match("^[a-zA-Z_]+$", module) is not None
-                if module in globals():
-                    continue
-                exec(f"import {module}", globals())
-
-        # Parse potential ModelConfig arguments
-        arguments = dict(**self.arguments)
-        arguments = traverse_and_apply(
-            arguments,
-            self._deserialise,
-            fn_kwargs={"trust": trust},
-        )
-
-        # Construct model based on arguments
-        return namespace_classes[self.class_name](**arguments)
 
     def _is_model_config_entry(self, entry: Dict[str, Any]) -> bool:
         """Check whether dictionary entry is a `ModelConfig`."""
@@ -225,6 +151,7 @@ class ModelConfig(BaseConfig):
         return {self.__class__.__name__: config_dict}
 
 
+# @TODO: Rename
 def save_config(init_fn: Callable) -> Callable:
     """Save the arguments to `__init__` functions as a member `ModelConfig`."""
 
@@ -260,69 +187,3 @@ def save_config(init_fn: Callable) -> Callable:
         return ret
 
     return wrapper
-
-
-def list_all_submodules(*packages: types.ModuleType) -> List[types.ModuleType]:
-    """List all submodules in `packages` recursively."""
-    # Resolve one or more packages
-    if len(packages) > 1:
-        return list(
-            itertools.chain.from_iterable(map(list_all_submodules, packages))
-        )
-    else:
-        assert len(packages) == 1, "No packages specified"
-        package = packages[0]
-
-    submodules: List[types.ModuleType] = []
-    for _, module_name, is_pkg in pkgutil.walk_packages(
-        package.__path__, package.__name__ + "."
-    ):
-        module = __import__(module_name, fromlist="dummylist")
-        submodules.append(module)
-        if is_pkg:
-            submodules.extend(list_all_submodules(module))
-
-    return submodules
-
-
-def get_all_grapnet_classes(*packages: types.ModuleType) -> Dict[str, type]:
-    """List all grapnet classes in `packages`."""
-    submodules = list_all_submodules(*packages)
-    classes: Dict[str, type] = {}
-    for submodule in submodules:
-        new_classes = get_graphnet_classes(submodule)
-        for key in new_classes:
-            if key in classes and classes[key] != new_classes[key]:
-                logger.warning(
-                    f"Class {key} found in both {classes[key]} and "
-                    f"{new_classes[key]}. Keeping first instance. "
-                    "Consider renaming."
-                )
-        classes.update(new_classes)
-
-    return classes
-
-
-def is_graphnet_module(obj: types.ModuleType) -> bool:
-    """Return whether `obj` is a module in graphnet."""
-    return isinstance(obj, types.ModuleType) and obj.__name__.startswith(
-        "graphnet."
-    )
-
-
-def is_graphnet_class(obj: type) -> bool:
-    """Return whether `obj` is a class in graphnet."""
-    return isinstance(obj, type) and obj.__module__.startswith("graphnet.")
-
-
-def get_graphnet_classes(module: types.ModuleType) -> Dict[str, type]:
-    """Return a lookup of all graphnet class names in `module`."""
-    if not is_graphnet_module(module):
-        logger.info(f"{module} is not a graphnet module")
-        return {}
-    classes = {
-        key: val
-        for key, val in module.__dict__.items()
-        if is_graphnet_class(val)
-    }
-    return classes
